feat: json working
